fixed makefile and added two functions\n strlcpy, strchr and strrchr still give problems

+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: maizpuru <maizpuru@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/26 18:39:30 by maizpuru          #+#    #+#             */
-/*   Updated: 2023/09/26 19:17:38 by maizpuru         ###   ########.fr       */
+/*   Updated: 2023/09/27 12:15:18 by maizpuru         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (!src || !dst)
